chore: rename misleading parse-args to parse-node-args
