+++ packages/webpack-cli/bin/cli.js
@@ -3,7 +3,7 @@
 'use strict';
 require('v8-compile-cache');
 const importLocal = require('import-local');
-const parseArgs = require('../lib/utils/parse-args');
+const parseNodeArgs = require('../lib/utils/parse-node-args');
 const runner = require('../lib/runner');
 
 // Prefer the local installation of webpack-cli
@@ -13,6 +13,6 @@ if (importLocal(__filename)) {
 process.title = 'webpack';
 
 const [, , ...rawArgs] = process.argv;
-const { cliArgs, nodeArgs } = parseArgs(rawArgs);
+const { cliArgs, nodeArgs } = parseNodeArgs(rawArgs);
 
 runner(nodeArgs, cliArgs);
